Add script, compose, ignore, readme and licence
We try to be minimalistic in terms of dependencies and implementation.
Our only direct dependencies are WebDriver library (Selenium) and
one-time password library (ROTP). Script implementation avoids
abstraction and indirection. However, we use wait helper from Selenium
to wait for the browser to finish loading or animations. We also use
Ruby exit hooks to log out gracefully from Mailchimp and close WebDriver
connection in case of error.

We tested the script on Ruby 2.6 as well as Ruby 3.2 with Firefox 109
and geckodriver 0.32.1.

We implement the script to go through the following steps:

1. Open Mailchimp login page to log in to billing history page.
2. Close cookie consent banner. Also useful as a check all the important
   scripts are loaded and initialized.
3. Submit username and password.
4. Submit one-time password.
5. Find an invoice issued in given month (assumes there is only one
   invoice per month).
6. Open the invoice as the top frame and remove interactive elements.
7. Save as PDF.

To do this, we need at least the following:

* username
* password
* secret to generate second-factor one-time password (two-factor
  authentication with authenticator app on Mailchimp is a requirement)
* web browser with WebDriver interface (for example Firefox with
  geckodriver)

The script can drive the browser locally (default) or remotely, and it
can optionally do so in headless mode. This is useful when executed
via containers.

We can pass secrets (username, password and two-factor authentication
secret) directly as environment variables, or we can store them in files
and reference them via equivalent environment variables with “_FILE”
suffix. This is useful in shared environments in which environment
variable values cannot be considered safe.

We also provide Compose file. It simplified integration between the
script, WebDriver interface and the browser. We run the script as
“archive” service in Alpine container with Ruby runtime from Docker Hub.
We run WebDriver interface and browser as “firefox” service in
standalone Selenium container for Firefox. We configure the script so
that it drives the browser from “firefox” service. Moreover, the
“archive” service running the script waits for “firefox” service to be
ready (via container healtcheck).

We do not use secrets in Compose file because Docker Compose (the most
popular implementation of Compose spec) does not support them without
setting up swarm mode (a cluster of Docker Engines) which is something
most people do not do. We believe people using Compose as-is do so on
their personal machines where environment variable values can be
considered safe.

We distribute the script under copyleft license (GPL-3.0).